+++ CHANGELOG.md
@@ -34,6 +34,10 @@ All notable changes are documented in this file using the [Keep a CHANGELOG](htt
 * Fixed [#148](https://github.com/NeoVintageous/NeoVintageous/issues/148): `:$` does not go to last line
 * Fixed [#87](https://github.com/NeoVintageous/NeoVintageous/issues/87): Double front slash doesn't escape properly
 
+## Removed
+
+* Removed: Deprecated SublimeLinter API calls
+
 ## 1.5.3 - 2018-03-24
 
 ### Fixed